More tests for data model compilation (#240)
* Tests for data model

* Debug log

* log statement for stalled out stop message, reduce timeout

* Try coverage with updates

* https instead of git

* disable sudo required

* Splitting up a test file

* Try gracefully shutting down during tests

* disable two tests

* Add exit code to coverage script
