+++ pubspec.yaml
@@ -23,7 +23,10 @@ dependencies:
 dev_dependencies:
   test: any
   http: ">=0.11.3+7 <0.12.0"
-  coverage: ^0.9.0
+  coverage:
+    git:
+      url: https://github.com/joeconwaystk/coverage.git
+      ref: master
 
 executables:
   aqueduct: aqueduct
