+++ test/command/generate_test.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:aqueduct/aqueduct.dart';
 import 'package:test/test.dart';
 import 'cli_helpers.dart';
-import 'generate_helpers.dart';
 
 void main() {
   group("Generating migration files", () {
