+++ test/command/migration_execution_test.dart
@@ -1,9 +1,7 @@
 import 'package:test/test.dart';
 import 'package:aqueduct/aqueduct.dart';
-import 'package:aqueduct/executable.dart';
 import 'dart:async';
 import 'dart:io';
-import 'package:args/args.dart';
 import 'cli_helpers.dart';
 import 'generate_helpers.dart';
